classifier/pyproject.toml
[project]
name = "filethat"
version = "0.1.0"
description = "Automated administrative document classification pipeline powered by LLMs"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.28.1",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0.2",
    "tenacity>=9.0.0",
    "rapidfuzz>=3.10.1",
    "python-json-logger>=3.2.1",
    "openai>=2.32.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.4",
    "mypy>=1.13.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "types-pyyaml>=6.0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/filethat"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "B", "UP", "SIM", "RUF", "BLE"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"
exclude = ["tests/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=filethat.normalizer --cov=filethat.referential --cov=filethat.schemas --cov-report=term-missing --cov-fail-under=60"
